react-flask-app/backend_drivers/requirements.txt
absl-py==0.10.0
altgraph==0.17
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
astroid==2.4.1
astunparse==1.6.3
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2
cssselect==1.1.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
-e git+https://github.com/pallets/flask.git@5b1d47f7e45ca6304009156882fcb6e4ebe55fbc#egg=Flask
future==0.18.2
gast==0.3.3
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
hyperlink==19.0.0
identify==1.5.6
idna==2.9
incremental==17.5.0
inexactsearch==1.0.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.4
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lxml==4.5.0
macholib==1.14
Markdown==3.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
nodeenv==1.5.0
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.4
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.6.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.2
PyHamcrest==2.0.2
pyinstaller-hooks-contrib==2020.9
pylint==2.5.2
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pyspellchecker==0.5.5
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
queuelib==1.5.0
regex==2020.10.15
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
Send2Trash==1.5.0
service-identity==18.1.0
silpa-common==0.3
six==1.14.0
soundex==1.1.3
soupsieve==2.0.1
spellchecker==0.4
sqlparse==0.3.1
srsly==1.0.2
starkbank-ecdsa==1.1.0
stevedore==1.32.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
tweepy==3.9.0
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.24.3
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.22.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
gunicorn==20.0.4
keras==2.4.3